+++ include/dom/html/html_input_element.h
@@ -67,10 +67,10 @@ dom_exception dom_html_input_element_set_disabled(
 	dom_html_input_element *input, bool disabled);
 
 dom_exception dom_html_input_element_get_max_length(
-	dom_html_input_element *input, long *max_length);
+	dom_html_input_element *input, int32_t *max_length);
 
 dom_exception dom_html_input_element_set_max_length(
-	dom_html_input_element *input, unsigned long max_length);
+	dom_html_input_element *input, uint32_t max_length);
 
 dom_exception dom_html_input_element_get_name(
 	dom_html_input_element *input, dom_string **name);
@@ -97,10 +97,10 @@ dom_exception dom_html_input_element_set_src(
 	dom_html_input_element *input, dom_string *src);
 
 dom_exception dom_html_input_element_get_tab_index(
-	dom_html_input_element *input, long *tab_index);
+	dom_html_input_element *input, int32_t *tab_index);
 
 dom_exception dom_html_input_element_set_tab_index(
-	dom_html_input_element *input, unsigned long tab_index);
+	dom_html_input_element *input, uint32_t tab_index);
 
 dom_exception dom_html_input_element_get_type(
 	dom_html_input_element *input, dom_string **type);
