+++ include/dom/html/html_select_element.h
@@ -18,13 +18,13 @@ struct dom_html_options_collection;
 struct dom_html_element;
 
 dom_exception dom_html_select_element_get_selected_index(
-		dom_html_select_element *ele, unsigned long *index);
+		dom_html_select_element *ele, uint32_t *index);
 dom_exception dom_html_select_element_set_selected_index(
-		dom_html_select_element *ele, unsigned long index);
+		dom_html_select_element *ele, uint32_t index);
 dom_exception dom_html_select_element_get_length(
-		dom_html_select_element *ele, unsigned long *len);
+		dom_html_select_element *ele, uint32_t *len);
 dom_exception dom_html_select_element_set_length(
-		dom_html_select_element *ele, unsigned long len);
+		dom_html_select_element *ele, uint32_t len);
 dom_exception dom_html_select_element_get_options(
 		dom_html_select_element *ele,
 		struct dom_html_options_collection **col);
@@ -37,13 +37,13 @@ dom_exception dom_html_select_element_get_multiple(
 dom_exception dom_html_select_element_set_multiple(
 		dom_html_select_element *ele, bool multiple);
 dom_exception dom_html_select_element_get_size(
-		dom_html_select_element *ele, unsigned long *size);
+		dom_html_select_element *ele, uint32_t *size);
 dom_exception dom_html_select_element_set_size(
-		dom_html_select_element *ele, unsigned long size);
+		dom_html_select_element *ele, uint32_t size);
 dom_exception dom_html_select_element_get_tab_index(
-		dom_html_select_element *ele, unsigned long *tab_index);
+		dom_html_select_element *ele, uint32_t *tab_index);
 dom_exception dom_html_select_element_set_tab_index(
-		dom_html_select_element *ele, unsigned long tab_index);
+		dom_html_select_element *ele, uint32_t tab_index);
 
 /* Functions */
 dom_exception dom_html_select_element_add(struct dom_html_element *ele,
